# Using 128 threads by default for cuda kernels