--- a/src/tensor_ops/select_and_gather/cuda_kernel.rs
+++ b/src/tensor_ops/select_and_gather/cuda_kernel.rs
@@ -1,8 +1,8 @@
 use crate::{
     shapes::{RemoveDimTo, ReplaceDimTo, Shape},
-    tensor::{Cuda, Tensor},
+    tensor::{launch_cfg, Cuda, Tensor},
 };
-use cudarc::driver::{DeviceSlice, LaunchAsync, LaunchConfig};
+use cudarc::driver::{DeviceSlice, LaunchAsync};
 
 const GATHER_PTX_SRC: &str = include_str!(concat!(env!("OUT_DIR"), "/gather.ptx"));
 const SELECT_PTX_SRC: &str = include_str!(concat!(env!("OUT_DIR"), "/select.ptx"));
@@ -36,7 +36,7 @@ macro_rules! impl_cuda_kernels {
                 let idx_strides = self.dev.htod_copy(idx.strides.into())?;
 
                 let fwd_fn = self.dev.get_func($GatherMod, $GatherFwd).unwrap();
-                let cfg = LaunchConfig::for_num_elems(numel as u32);
+                let cfg = launch_cfg(numel as u32);
                 let params = (
                     numel,             // const size_t numel,
                     inp.data.as_ref(), // const float *inp,
@@ -74,7 +74,7 @@ macro_rules! impl_cuda_kernels {
                 let inp_strides = self.dev.htod_copy(inp.strides.into())?;
                 let idx_strides = self.dev.htod_copy(idx.strides.into())?;
 
-                let cfg = LaunchConfig::for_num_elems(numel as u32);
+                let cfg = launch_cfg(numel as u32);
                 let params = (
                     numel,             // const size_t numel,
                     grad_inp,          // float *grad_inp,
@@ -122,7 +122,7 @@ macro_rules! impl_cuda_kernels {
                 let dst_strides = self.dev.htod_copy(dst.strides().into())?;
 
                 let fwd_fn = self.dev.get_func($SelectMod, $SelectFwd).unwrap();
-                let cfg = LaunchConfig::for_num_elems(numel as u32);
+                let cfg = launch_cfg(numel as u32);
                 let params = (
                     numel,             // const size_t numel,
                     inp.data.as_ref(), // const float *inp,
@@ -163,7 +163,7 @@ macro_rules! impl_cuda_kernels {
                 let idx_strides = self.dev.htod_copy(idx.strides.into())?;
                 let out_strides = self.dev.htod_copy(out.strides.into())?;
 
-                let cfg = LaunchConfig::for_num_elems(numel as u32);
+                let cfg = launch_cfg(numel as u32);
                 let params = (
                     numel,             // const size_t numel,
                     grad_inp,          // float *grad_inp,