Resolves #526 

This shaves off a bit of time during forward & backward pass by allowing cuda to split operations across multiple groups.

Shaves off a millisecond from `conv2d` bench, and when benchmarking with https://github.com/coreylowman/image-classification on batch size 64, the forward time goes from approximately 36ms -> 28ms, and backward from 60ms -> 51ms